Weather API is integrated with my rest controller get request
