+++ mocktest/src/main/java/com/mockitotest/MockTest/Repository/ItemRepository.java
@@ -4,7 +4,7 @@ import com.mockitotest.MockTest.Entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
+//@Repository
 public interface ItemRepository extends JpaRepository<Item, Long> {
     public Item findByNameIgnoreCase(String name);
 }
