+++ weather-api/.gitattributes
