+++ weather-api/pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>weather</artifactId>
+	<artifactId>weather-api</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>weather</name>
 	<description>Weather API project</description>
