+++ weather-api/src/main/java/com/weather/model/HourlyUnit.java
