+++ weather-api/src/main/resources/application.properties
